Added MoveWheels.CS
This is a script that, when applied to an object with four hinge joints
and wheels, will cause the wheels to rotate (making the car move). At
this point, it will only work under specific circumstances, but it can
be modified to work with script-generated robots. This is a work in
progress.