Created a basic wheel motion/animation function class.
Contains a number of functions useful for making the wheels of a model
move. This includes a basic setControls function, and a function that
allows you to move an objects pivot point to the center of the object.
It does not actually do anything now (that is, it is not used in the
init script), but its a good first step.

I have also moved MoveWheels.cs to the closet, because it has become
outdated and will likely remain unused.