Radius now calculated from vertices of mesh.
Added xml comments to related functions.  The mesh is constructed with a
tolerance of half a centimeter, easy to adjust with needs of Unity's
wheel collider.  Now uses Box's extend function to include out of bounds
points.  Next step is to compare dimensions of box to find the radius
axis.  Also need to look into transforming vertex coordinates to
assembly axes.