The childNormal of the rotation joint for wheels is now used to calculate radius.  Vertex traversal is now down in multiple threads.
The code looks at the childNormal.  If it lines up with a vector, the
other two dimensions of the bounding box are averaged and then halved to
determine the radius of the wheel.  Multiple threads are now used for
SubOccurrences, whilst Occurrences in the group are still handled
seperately.  Sub-subOccurences may break things.  Sorry for the large
commit, meant to do one with just the childNormal stuff but forgot.