Temporary Drive code and Error Handling
I created a file called 'FakeInit.cs' which is meant to contain some
simple drive code that other members of the Unity team can access (it is
by no means a permanent thing, however, because it may change or may not
be used at all because the user will upload there own drivetrain code).
I also created a file called ErrorStuff.cs. It, like its name, is very
simple and is meant to hold any custom exceptions the Unity team wants
to make. At the moment, it has a PWMAssignedException, which is meant to
signify that a PWM port that the user attempts to sign has already been
used already (see the PWM port dictionary in Controls.cs). I did have a
problem, however, with one of the wheels. I don't know if its the wheel
collider, but one of the configurable joints rotates a wheel backwards.
I assume it was a bug and that it will be fixed in the coming updates. I
am 50% confident that the wheel collider was not reversed, however. I
also updated Controls.cs to make FakeInit work correctly.

Note: You can always use the "old reliable" drivetrain code if the
equation method does not work.