Actual Shader Commit
I noticed that I made a couple of mistakes in my last commit so I
thought I would try this again.

Uploaded a new VertexColor shader that includes basic vertex shading and
edited UnityRigidNode to use that instead of the "Diffuse" shader.
Double wheel collider creation was also commented out because we are
trying a different strategy. FakeInit was moved to the closet.