Quick Fix
There are no Tuples in .NET 3.5 (and I don't believe that Unity
implemented its own form of Tuple) so I had to create a class to imitate
the basic function of tuples. I am uploading this now because I forgot
to before. This file also contains a class called JointContainer, which
I originally wanted to use to store data. I realize that the folder is
getting messy, and that is why i want to restructure everything.