Revamp
Cleaned-up all of my code so it (hopefully) only uses what is needed. It
was also re-organized, and unnecessary files/classes were removed. It
may need some testing, but it is much less convoluted.

What I still need to do: match the Unity Torque curve with that of a
standard VEX motor.