SetMotor Update and Minor Fixes
Updated the SetMotor function so that it explicitly sets the target
angular velocity of the configurable joints (that operate the wheel
animations) to 0 if the PWM signal is 0. I also took out the multiplier
that made robots wheelie every time they drive.

I also pushed some minor fixes that got everything working with the new
API, because I feel that it's getting quite annoying to have make these
fixes every time we come back to the master branch.