Minor Changes
I am not sure if the conversion factor needs a 100f multiplier. Logic
says that it does not, but someone on the forums did note that the
values may need to be fudged a little until they "feel right".

Other than that, I modified the piston management function a bit to stop
and reset a piston's stopwatch if it continuously accelerates for over 1
second (which it should not). I still need to make it reset once the
piston hits a limit.