New field features. Camera bugs fixed.
The field operations now run more efficiently than they did before.
"EnableCollisionObjects" was changed to "AddCollisionObjects" and a new
method called "SetCollisionObjectsEnabled" allows you to enable and
disable collision for the objects specified. The camera now resets when
the robot is reoriented. Mouse controls for panning in FPV mode were
changed to the 'W' and 'S' keys for panning up and down. An "Extensions"
class was also created to implement an "AddAll" method for the
"Dictionary" class.