Improved collision; Less buggy and more efficient.
Before this commit, the field had an issue where the ground mesh was so
thin that if the robot could fall through the ground if spawned ~10 ft
in the air. The wheels also got stuck in the ground often. To fix this,
I added functionality for manually creating collision boxes for the
field. I replaced the majority of mesh colliders with collision boxes to
increase performance, and only kept collision meshes where they were
absolutely necessary (the scoring platforms and tote chutes).