Tote functionality added. Field physics improved.
By setting the "isKinematic" property for the field to true, I reduced
the amount that the wheels sink into the ground. I have also added the
tote class, which, well, represents a tote... and I added field
functionality so that you can have different physical properties for
different field components.