Started on field exporter.
I have begun to create an application that exports field files which can
be imported into the simulator. So far, I have a GUI that allows the
user to scan the components of an assembly file and select which
components have certain properties by dragging and dropping them into
various property windows.