Deleted field switcher. Camera improved.
The field switcher has a strange inconsistent bug that is sometimes not
repeatable.  Because of this and the fact that the field switcher will
be completely different after the field exporter implementation, the
field switcher has been removed for now. The camera has also been
improved so that it will remain in the same state after the robot
reloads or is reoriented.


Former-commit-id: 348e578d624212d34b69b38f226a43c3c563338c