GUI revamped. Basic export functionality.
I have completely revamped the GUI of the field exporter. There is now a
tab system for switching between various field properties (right now
there is only "Collision" and "Export", though). In the collision tab,
there is a collision objects treeview that contains any specified
objects that will have collision properties. To add collision components
to the view, click "Select in Inventor" and click the components you
want to add. Then, click "Add Selection" to add the components or
"Cancel Selection" to deselect everything and exit select mode. When
adding components, a progress window will appear with a progress bar and
a label describing which component is currently being added. After the
adding is completed, a collection of treenodes will appear in the
collision objects treeview. When you aren't in select mode, you can
hover the mouse over various nodes, and their corresponding components
will light up in Inventor. Conversely, in select mode, you can hover the
mouse over Inventor components, and their corresponding treenodes will
light up. This makes it easy to see if you have already added a
component. It doesn't matter if you add a component twice, though; the
collision objects treeview will ensure that there are no duplicate
nodes. It will also add nodes intelligently; if a part is added that is
a child of a subassembly, the subassembly will be added also. In other
words, only top-level assemblies can be root nodes. This will keep the
collision objects treeview as close as possible to the Inventor
treeview. In the export tab, you can click "Browse" to select the export
file path and then type a filename for the BXDA mesh file. Then, click
"Export." A progress window similar to the one I described earlier will
appear, this time showing a progress percentage label. When the process
completes, you will find a BXDA file with the specified name in the
specified location.