New export methods.
Instead of allowing the user to specify the file names of the BXDF and
BXDA files, the field exporter only allows the user to select the output
directory. The BXDF and BXDA files are then named description.bxdf and
mesh.bxda. This allows for a cleaner loading system in Synthesis.