Added awesome camera features.
Driver station view now looks towards the robot (not directly, but
enough so that you can see both the robot and most of the field at the
same time). A new camera mode, Freeroam, was added. Now you can explore
the entire scene with ease. Hold left click and move the mouse to travel
forward and backward and turn left and right. Hold right-click and move
the mouse to rotate your view in all directions. Hold both the left and
right mouse buttons and move the mouse to slide left, right, up, and
down. Use the scroll wheel to zoom in and out.