Fixed build errors. Stuff actually works now.
Title says it all. Well, almost all. There is a weird bug (in my version
at least) where the vertex normals are flipped in every mesh (including
the robot). The problem is a Unity issue for sure, because my working
version uses the same DLL and works fine. It's probably a project
settings thing, but IDK.