Minor design changes.
There is now a right-click option to change the name of a PhysicsGroup.
The color theme was changed to resemble Synthesis colors better. The
images for creating PhysicsGroups were removed. A Synthesis logo was
added to the background in the Export tab. The treeview now only selects
inventor components when a node is selected, not when it is hovered
over.