Wow, lots changed.
Improvements of all sorts took place, and it would take me a long time
to cover everything. But the major changes were stability improvements.
It is very very hard to crash the exporter now, whereas before you could
simply close inventor during an operation and crash everything. Now, the
moment connection is lost to inventor, a message box will pop up asking
the user to either reopen the assembly or close the exporter before
anything continues. This makes the exporter and the field assembly
appear to be linked together, making it hard to mess things up. Another
stability improvement was made so that if an exception IS thrown (again,
very unlikely from my testing), a dialog window will appear asking the
user to send their steps prior to the error occurring to the BXD team,
along with the details provided in a text box.