Huge stability improvements.
It is now very very very hard to crash the field exporter. There is a
lot of safety to ensure the user doesn't do anything he/she isn't
supposed to. There is also a field selector to initially define what
document the field exporter should reference. This eliminates the
possibility that the user will accidentally work with the wrong file. If
you are reading this commit message, you should also read the TODO in
Program.cs, because I have discovered something that could make both the
field exporter and the robot exporter way faster, more stable, and
easier to use.