Camera improvements.
The camera's up and down movement is more predictable in orbit mode.
Also, the camera's height does not change when the robot moves anymore;
it stays a consistent height above the robot until the user changes the
camera's position. The camera will still rotate horizontally around the
robot to keep the driver informed about what is going on in front of the
robot.