Made changes for BXDA compatibility system.
Made some changes to the BXDA file so there is now a GUID right after
the version number.  This is used to prevent users from simply replacing
existing BXDA files with different ones, unaware of the consequences.
For example, if a field's BXDA file is replaced by that of a different
field, Synthesis will alert the user that the mesh is incompatible.
Note that because the mesh layout has changed, old robot files will not
work.  But don't re-export anything yet because there will probably be
more updates to the BXDA file soon.