New BXDJ Writing Implemented; Started on writing.
The robot exported can now export the new XML-based BXDJ file.
Miscellaneous stability and design improvements were also made.  Changed
the versioning system a little bit; the third digit in the version
number is ignored by the parser and used simply for distinguishing
between different reader/writer versions, not file design versions.
Also fixed a bug with the field exporter in which an error would occur
when a non-part leaf occurrence of an assembly was exported (no
SurfaceBodies on a non-part).  Wrote an XSD for the BXDJ file.