User-friendly improvements.
No more process window - everything is now built right into the main
application window.  Instead of displaying a progress window for a
nanosecond when adding a selection from Inventor, the loading cursor
indicates that work is being done (it is usually virtually instantaneous
to add a selection from inventor, but if there is a large assembly and
you are adding the entire thing, it can take up to a second). The export
form also has its own built in progress bar now, and it is much nicer
looking than before (plus window overlay issues and cross-thread
operations aren't a problem anymore). There is also a help button that
links to the field exporter tutorial webpage and the export form has a
nice new "Inventor to Synthesis" background.