Design changes and added functionality.
No more PhysicsGroups... kind of. They are now called Property Sets
because physical properties are only one type of properties that can be
defined. The box collider was re-added, and a new collider, the sphere
collider, was implemented. I figured that having everything be a mesh
collider would slow down the export process unnecessarily (and it did).
With the box and sphere colliders, you can define the scale properties
in case you want the bounding box/radius to be, for example, twice as
large. The changes in functionality resulted in changes in the BXDF file
type, so version 2.2 was implemented. GUI improvements were also made.
More functionality to come.