Export only takes 30 seconds!!!
MAJOR improvements have been made to the export process.  Instead of
exporting each Inventor component individually as its own mesh, the
exporter refers to the part file that each component references and
duplicates that mesh for each identical occurrence. The
location/rotation properties that are component-specific are stored in
the BXDF file. This alone makes the export process 10 times faster. In
addition, the BXDA file for the 2015 game field, for example, only takes
up 3MB of space, vs. 85MB before the optimizations were made.  In the
process of making these optimizations, the BXDQuaternion object was
added to store each component's rotation properties for easy
implementation in Unity (euler angles don't seem to suit well with
Unity). A method was added in Utilities.cs to convert the
Inventor.Matrix object to the new BXDQuaternion object.