Unity4 - Made fixes and improvements on the main menu.
Changed 'Load Custom Robot/Field' to 'Choose Directory' so that users
can choose a default directory that would scan for robots/fields.

The Main Menu now remembers the last robot/field you loaded in with and
defaults the selection to that so that you don't have to go through the
trouble of selecting it again.