Fixed bug where robots aren't loaded correctly.
I have fixed a bug in the robot exporter in which when you load a robot,
there is a chance that the meshes of the robot will be mixed around.
This was the result of using .NET functions to get all .bxda files from
a folder, thus the files could be returned in an incorrect order. This
was fixed by loading the .bxdj file first, then using the file names
referenced there.