Added object tracking for replay, created FSM
Added a tracker component that stores transform information for its
parent object for a set number of past frames. The tracker keeps this
information in a fixed queue, which I created as a structure to quickly
add elements to a collection and overwrite the oldest element when a new
one is added to a full queue. I also created a finite state machine in
preparation for the replay analysis mode. An FSM will make the code a
whole lot cleaner and less bug prone, and it increases scalability if we
want to add more modes in the future.