Created replay feature, fixed FixedQueue bug.
The replay feature is now functional, and it allows the user to pause
the scene and phase through the past few seconds of the scene. You can
also resume the scene from any position in the replay. Also fixed a bug
in FixedQueue that was causing occasional innacurate values for a given
index, and a similar issue existed in the IEnumerator.