Added collision tracking, GUI
Added a feature to track collisions and display them visibly in the replay state. In doing this, I made the Trackers more efficient by making the frame detection happen once in the MainState instead of multiple times in each Tracker (this also simplified the Tracker class quite a bit). The GUI has a slider that can be scrubbed to adjust the time in the replay, and there are also play, pause, and rewind buttons. Collisions are displayed as white dots and are low opacity until they become relevant in the replay. You can click these dots to snap to the point in time that the collision happened. There are also keyframe markers on the slider that indicate collisions that you can snap to. There is a vertical slider that can be toggled and adjusted to set the collision impulse threshold (what is considered to be a notable collision by the replay state). There were quite a few other small changes but those are pretty insignificant and can be seen via the diff.


Former-commit-id: 5188d32174ed1693284ba3e06a1478baab78f483