Improved collision detection accuracy and added collision consolidation.
Improved collision detection accuracy by removing collision callbacks and instead iterating through the collision manifolds each frame. This also ensured that the collision tracking was synchronized with the transform tracking by the Trackers. I also added a feature that highlights the part of the robot that a collision is associated with when the mouse hovers over the collision dot. I added a collision consolidation feature, where if a single impact lasts for multiple frames and registers as multiple contacts, you can select the first contact and the rest of the contacts in the chain will be combined with the selected contact. This will be useful when exporting to an external program for FEA.