Improved collision tracking
Collision tracking is now fully integrated into the built in collision callback system. The primary reason for doing this was to avoid duplicating a GetManifoldByIndexInternal call, which generates garbage. That function is already being called in BPhysicsWorld.cs, which we don't control, so might as well use it's junk instead of making our own :neutral_face: