Fixed undefined behavior in collision tracking
I fixed the undefined behavior in collision tracking and finally found a solution to tracking in general that is efficient, accurate, and doesn't make the GC go mad. Essentially the new system has a callback listener (which can have an unlimited number of callbacks attached to it btw), which is linked to the physics tick. The trackers are linked with this tick instead of being updated manually from MainState. When it comes to manifold iteration, I've figrued out a way to override the BulletUnity collision callback system so it's synced with the physics loop instead of FixedUpdate() that way we don't have to worry about skipping collisions. So everything is good to go now!