Fixed rare tracking error
Fixed an issue where sometimes more frames would pass than the number of available spots were in the list of contact points. Now there is a null check, and empty contact points default to an empty list instead of a null value in case two or more frames pass and contact points are added retroactively.


Former-commit-id: d2b899bb1cc4752e7f7ae77e2e345ec0843c1b2b