Fixed rare tracking error
Fixed an issue where sometimes more frames would pass than the number of available spots were in the list of contact points. Now there is a null check, and empty contact points default to an empty list instead of a null value in case two or more frames pass and contact points are added retroactively.