Started cleaning up UI code with state machine
I've started cleaning up the monolithic MainMenu.cs by integrating the state machine with the menu system. This makes it very easy to add new windows and make changes to the UI, and also makes the code more organized and readable. This process is almost complete; the rest shouldn't be very difficult.


Former-commit-id: 414314d3c7c7455d7d5b277acdad1535a55accd0