Started on robot camera movement arrows integration
The robot camera can now be moved via the movement arrows. However, there are currently a few issues. Unintented behaviors occur when dragging the arrows and driving the robot at the same time, and the movement arrows are visible by the robot camera. These issues will be fixed in the next commit.


Former-commit-id: ab567463f32c7a1bdd9271f53aa9a88d3f43f9c9