Fixed a StateMachine issue and DPM bug
Fixed an issue with the state machine in which objects would be enabled when their state became active even if their parent state ended with them inactive. Now, the state machine will keep track of which objects should be reactivated. This also fixed an issue where the movement arrows would be visible in the replay mode.