Started on updating DPM physics
Started integrating updated DPM physics where the game elements actually behave in the physical world. Everything is working great EXCEPT for this annoying as heck bug where some undefined force gets applied to the game element when it is collected at high speeds. This will hopefully be fixed in the next commit.


Former-commit-id: 08fefe1427bada0611dae2b134c3ba5cf14977b4