Created a new filtering algorithm for network meshes.
Added a new algorithm for filtering transform data sent over the network to the client. This new system keeps track of two velocity values, one closely matching the velocity values sent by the server (with some small smoooth filtering), and one that adjusts for any offset between the client positional information and the server positional information. The filter on the second velocity value has both a very high response and a very quick recession. That way, small changes are almost unnoticeable on the client, but if there is any large offset (from a lag spike, etc.), then the client is quick to correct it.