Fixed issue where robots exported sideways don't drive properly.
Fixed the issue by calculating the best fit plane to approximate the chassis plane and take its normal to figure out which direction the wheels are facing. Then compared the centroid of the root node and calculated which side of the plane it was on in order to figure out which way to send the raycasts. Also did some other math to figure out which way is forward and right relative to the robot. The robot generation code itself is also a bit cleaner.


Former-commit-id: d82f20f7308e9175cac588c34f3b7a525bc527f3