Made some minor StateMachine changes
Changed the scene global warning to an error message since not having a global state machine is almost certainly catastrophic. Also made it so a State will never have a null StateMachine after it has been pushed. In other words, the StateMachine property in a State instance represents the last StateMachine that ran the State.


Former-commit-id: e9ede4a4a9b2ddfdd0e1649465fb4b7354b04976