Final Bug Fixes, Code Cleaned, Comments Added, Manager and Testing Functionality
Fixed the way the axis arrows stay facing the camera. Cleaned the code, got rid of unnecessary lines and functions, and added comments everywhere. Finally, added a GizmoManager class to make sure only one gizmo is spawned at a time. Added a temporary button on the home bar to demo the gizmo on a cube. Added press escape to exit gizmo.